move trajectory commands to autos package
