+++ src/main/java/frc/robot/RobotContainer.java
@@ -18,8 +18,6 @@ import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.ParallelDeadlineGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
 import edu.wpi.first.wpilibj2.command.WaitUntilCommand;
-import frc.robot.commands.drive.AutoFollowTrajectory;
-import frc.robot.commands.drive.AutoGoToPose;
 import frc.robot.commands.drive.DriveDistanceMeters;
 import frc.robot.commands.drive.TankDrive;
 import frc.robot.commands.feeder.FeederAcceptCargo;
@@ -50,6 +48,8 @@ import frc.robot.util.controllers.ButtonBox.ClimbBar;
 import frc.robot.util.controllers.ButtonBox.ClimbDirection;
 import frc.robot.util.ThisRobotTable;
 import frc.robot.commands.LimelightToggle;
+import frc.robot.commands.autos.AutoFollowTrajectory;
+import frc.robot.commands.autos.AutoGoToPose;
 import frc.robot.commands.autos.DriveWithWaypointsPlan;
 import frc.robot.commands.autos.TurretTrackWithGlobalPose;
 import frc.robot.commands.autos.SetGlobalPoseToWaypoint;
