+++ src/main/java/frc/robot/commands/autos/AutoGoToPose.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.drive;
+package frc.robot.commands.autos;
 
 import java.util.ArrayList;
